pyproject.toml
[project]
name = "cuba-payment-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Jose Miguel",email = "jmperezperez25@gmail.com"},
    {name = "Yasdiel Alejandro",email = "yasdielalejandro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.5,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [
    {include = "cuba_payment_back"}
]

[tool.poetry.scripts]
start = "cuba_payment_back.cli:start"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
